Update twitter login code.
The Twitter handle is now stored in the email column, which just makes more sense, as it is as unique as an email. The middleware returns more information, including
some twitter activity stats, the twitter *name*, and twitter image URLs.

+++ app/models/identity.rb
@@ -26,6 +26,8 @@ class Identity < ActiveRecord::Base
   private
   
   def delete_user_if_deleted_last_identity
+    return unless user
+
     return if user.identities.any? { |identity| identity.id != self.id }
     user.destroy
   end
